+++ tests/test_pipe.py
@@ -3,8 +3,8 @@ from io import StringIO
 
 import pytest
 
-from pipe_utils import Pipe, it, map_
-from pipe_utils.pipe import Catch, Then, obj
+from pipe_utils import Pipe, as_list, it, iter_reverse, map_
+from pipe_utils.pipe import Catch, P, Then, obj, unwrap
 
 
 def test_pipe_yields_data():
@@ -32,23 +32,23 @@ def test_pipe_or():
     assert (Pipe(5) | str).get() == "5"
     assert (Pipe([3, 2, 1]) | sorted).get() == [1, 2, 3]
     assert (
-                   Pipe("Hello")
-                   | list
-                   | (lambda data: map(lambda c: c * 2, data))
-                   | "".join
+                 Pipe("Hello")
+                 | list
+                 | (lambda data: map(lambda c: c * 2, data))
+                 | "".join
            ).get() == "HHeelllloo"
     assert (
-                   Pipe(1)
-                   | (operator.add, 5)
-                   | (operator.mul, 2)
+                 Pipe(1)
+                 | (operator.add, 5)
+                 | (operator.mul, 2)
            ).get() == 12
 
 
 def test_err_is_caught():
     pipe = (
-            Pipe([-1, 0, 1])
-            | (lambda data: map(lambda x: 1 / x, data))
-            | list
+          Pipe([-1, 0, 1])
+          | (lambda data: map(lambda x: 1 / x, data))
+          | list
     )
     assert pipe.get_or_default(-1) == -1
 
@@ -86,10 +86,10 @@ def test_catch():
 
 def test_pipe_err_state_is_chained():
     assert (
-                   Pipe(1)
-                   | (operator.truediv, 0)
-                   | (operator.add, 1)
-                   | Then(print, end="")
+                 Pipe(1)
+                 | (operator.truediv, 0)
+                 | (operator.add, 1)
+                 | Then(print, end="")
            ).get_or_default("Oops!") == "Oops!"
 
 
@@ -122,17 +122,17 @@ def test_get_or_raise_ignores_wrong_exception_type():
 
 def test_pipe_then_and_catch_classes():
     assert (
-                   Pipe([-1, 0, 1])
-                   | Then(lambda data: map(lambda x: 1 / x, data))
-                   | Then(list)
-                   | Catch(ZeroDivisionError, lambda _: [0])
+                 Pipe([-1, 0, 1])
+                 | Then(lambda data: map(lambda x: 1 / x, data))
+                 | Then(list)
+                 | Catch(ZeroDivisionError, lambda _: [0])
            ).get() == [0]
     with pytest.raises(ZeroDivisionError):
         (
-                Pipe([-1, 0, 1])
-                | Then(lambda data: map(lambda x: 1 / x, data))
-                | Then(list)
-                | Catch(TypeError, lambda _: [0])
+              Pipe([-1, 0, 1])
+              | Then(lambda data: map(lambda x: 1 / x, data))
+              | Then(list)
+              | Catch(TypeError, lambda _: [0])
         ).get()
 
 
@@ -161,3 +161,8 @@ def test_get_or_raise_controls_error_chaining():
 
 def test_obj():
     assert obj.upper()("hello") == "HELLO"
+
+
+def test_pipe_unwrap_returns_piped_data():
+    assert P >> "hello" | unwrap == "hello"
+    assert P >> [1, 2, 3] | iter_reverse | as_list | unwrap == [3, 2, 1]